// src/app/main/model/search/filter-model.ts
export interface FilterModel {
    inquiryID?: any;
    commodity?: any;
    inquiryType?: any;
    client?: any;
    pol?: any;
    pod?: any;
    createdBy?: any;
    sentTo?: any;
    termofService?: any;
    dateCreate?: any;
}